Adding .Net Driver Facade pattern
Adding a facade pattern to the use of the driver in testkitbackend. This is to enable
the swapping in/out of .net driver for cbindings driver.

Added new project that contains the driver facade and calls into .net driver.
Used the facade in NewDriver, CloseDriver and NewSession as proof of concept.